Added Credits and Instructions
