+++ bwj-game/Assets/Scenes/Credits.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 511c5563249d0cf498eabe30ba45e23c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
