+++ bwj-game/Assets/Scenes/Instructions.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0475a7ea34a08b6409763b5ade821370
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
