+++ bwj-game/Assets/Scripts/Menu.cs
@@ -1,0 +1,33 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class Menu : MonoBehaviour {
+
+    public void PlayGame ()
+    {
+        SceneManager.LoadScene("CHASE_BaseV2");
+    }
+
+    public void QuitGame ()
+    {
+        Debug.Log("QUIT!");
+        Application.Quit();
+    }
+
+    public void Back ()
+    {
+        SceneManager.LoadScene("Menu");
+    }
+
+    public void Credits ()
+    {
+        SceneManager.LoadScene("Credits");
+    }
+
+    public void Instructions()
+    {
+        SceneManager.LoadScene("Instructions");
+    }
+}
